requirements.txt
autopep8==1.3.5
certifi==2018.8.13
chardet==3.0.4
Django==2.0.5
django-debug-toolbar==1.9.1
django-silk==3.0.1
ExifRead==2.1.2
ffmpy==0.2.2
gprof2dot==2016.10.13
idna==2.7
Jinja2==2.10
lxml==4.2.4
MarkupSafe==1.0
mysqlclient==1.3.13
netdisco==2.0.0
netifaces==0.10.7
numpy==1.15.0
Pillow==5.2.0
pycodestyle==2.4.0
Pygments==2.2.0
python-dateutil==2.7.3
pytz==2018.5
raven==6.9.0
requests==2.19.1
six==1.11.0
sqlparse==0.2.4
uPnPClient==0.0.8
urllib3==1.23
uWSGI==2.0.17.1
zeroconf==0.20.0